Add unit tests for edit, delet all, update status and dom manipulation funtions
